hoa/inspire12-order (#4)
* feat(order): order 필요한 데이터 추가, Exception 정의

* feat(order): order 필요한 데이터 추가, Exception 정의

+++ order/src/main/java/com/smart/commerce/order/module/order/infrastructure/repository/OrderJpaRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Component;
 
 // TODO JPA 연동 이후 변경 예정
 @Component
-public interface OrderJpaRepository extends JpaRepository<OrderEntity, Integer> {
+public interface OrderJpaRepository extends JpaRepository<OrderEntity, Long> {
 
 }
